Fix indexing issues with fakeRows in KisTagModel
Before this commit, index of a tag was different depending on
whether it was fetched in tagForIndex() or data() function
(data would subtract it with fakeRowsCount and tagForIndex
would just use raw index) which would cause indexing issues
including crash on checking "All" tag.
This commit adds subtracting by the number of fake rows
in tagForIndex(), which fixes the indexing issues.