Fix TagFilter when filtering by "All"
Before this commit, TagFilter would always check resources
if they have a tag specified in the combobox. That is not
desired in case of tag "All" which is a fake tag that
should contain all resources without the need to add
all of them to it.
This commit adds a special rule for "All" tag to make sure
all resources are captured.