Disable removing and renaming special tags
Before this commit, "All" tag would be easily renamed or removed
as user wished. This commit adds a way to recognize special
read-only tags by checking their id; if it's negative, it means
the tag is not from the original database, but added in KisTagModel
automatically for convenience.
All tag is recognized by id instead of url, because there might be
more special tags: for example "All untagged" or "All textured"
or whatever users will come up with and ask for.