Save filtered presets to the current tag
Before this commit, the button to save the filtered presets
to the current tag would do nothing. This commit makes sure
that the resources previously added to the tag are removed
from the tag, and the new set of resources is saved into the tag.

Right now fitlering works in such a way that it only displays
resources that has the tag mentioned in the combobox,
so this method can only remove some presets from the tag,
but not add them.
It needs some usage discussions. There might be a new button
needed "Save to a new tag" as well.