Remove regexes from KisResourceSearchBoxFilter
Regular expressions to check for the beginning and the end
of the filter part were both unnecessary (since it can easily be done
using startWith and endsWith) and broken for new brushes,
for example "f)_Bristle-4_Glaze" wouldn't match the exactMatch
regular expression because of symbols inside.
Moreover it would not allow for any non-latin characters or
sumbols.
This commit changes those expressions to the usage of startsWith
and endsWith.