Increase the limit in Layer Offset to 100k
Before this commit, there was no maximum set on the double
spin box which holded the value. Because by default it would
be filled with half of the width and height of the layer
extent, there will be some maximum value set bigger
than 99 which is the default for double spin boxes.
This commit fixes this behaviour by setting 100 000
as a maximum explicitely.

BUG:414625