Fix 1px brush offset in line tool
This adjust the diferent position of KisPaintInformation
from a line to be inside a line between starting and
ending points. This is particularly important as brushes
that use 1px perfect pixels require the line to be perfect
in order to avoid looking jagged. In 1px size with
sharpen active we only use the start and end
KisPaintInformation points.

This patch also adjusts the outine on 1px brushes to
adjust to the grid to better preview where the paint is
going to be applyied

BUG:407405