Add QCompleter to tag filter box
Before this commit, there was no completion in the resources
filter text box. After this commit it completes tags.
Unfortunately it's not compatible with syntax of the filter box
(where tags should be written in []).
Now artists need to first use the completion and then add [] around
it, and it still won't work because filtering by tags is not coded yet.