Add a new role to get tag and resource from models
Before this commit, there was no way to get a proper KisTagSP
or KoResourceSP without the appropriate model. Problem was
for example in case of the brush presets docker, which had
two ProxyFilterModels, so in the second one it would have to ask
for the source model twice to get to the appropriate source model.
Those functions are there for usability for the developer.