Fix exiv2 compilation on Windows
Before this commit, 'make' would be used instead of the actual
make executable (on my sysem called 'mingw64-make.exe'). This commit
changes the call to $(MAKE) instead which deals with that issue.

Before this commit, exiv2 would not compile because of constness
issues. This commit adds a patch for it with const casting.