Disable post-install step in ilmbase
Before this commit, some manual copying of files was performed.
It both doesn't seem to be needed and breaks the compilation.
This commit comments the commands out (OpenEXR needs to be
fixed and reevaluated since it looks like it provides its own
ilmbase, but since for now OpenEXR doesn't work, commands are
just commented out in this commit instead of removed).