Make Add subbrush off on changing multibrush tool's type from Copy Translate
This commit makes the Add option of subbrush off when changing from
Copy Translate type to any other type. The option is toggled off
even for Copy Translate type of multibrush. Previously the Add
option stayed on and drawing with other tools was not possible unless
it was turned off.

BUG:415651