ORA export, write entire layers instead of cropping them
This change stores the full image layer data in exported ORA files, even when part of the layer is outside of the visible canvas.
(similar to how layer data is stored in KRA format) Properly adds the "x" and "y" attribute to the resulting layer stack to re assemble
 the image. This change address a previous issue where the layer data would become cropped when exporting to ORA format, resulting in
lost layer data in some cases.