Make color selector aware of UI scaling
Qt virtualizes the coordinate system and keeps it independent
of (Hi-DPI) UI scaling, that helps to not break applications
not aware of this feature, but automatically upscaled images
of course are not very nice especially with fractional factors.

This commit tries to properly determine the true device pixel amounts
and renders the selector at this resolution.

Note: fractional scaling needs thorough testing with Qt 5.14,
earlier versions are know to still have many rendering glitches.
Only tested with Qt 5.12 and QT_SCALE_FACTOR environment variable so far.