Initialize transform tool args when ending stroke
Before this commit, transform tool would get an
assert when switching to another tool without
any changes.
This commit removes trying to cancel the stroke
in the finishStrokeCallback() call
and initializes m_savedTransformArgs so the finishing
can happen as always.

WARNING: it probably isn't the "proper" solution
for this issue. It does work correctly, but it adds
an unnecessary undo step in the situation when cancelling
should be used instead of finishing the stroke
(for example when the user switches to transform tool
and then immediately to another tool without making any changes).

CCBUG:414672
CCBUG:415625
(cherry picked from commit 8ef04427c4b8e63265ab4bb3456659cf4b0d39a4)