Reset TagsResourcesModel after TagM or ResourceM
Before this commit, no communication happened between
all of those three models. When a Tag or a Resource
was removed, TagsResourcesModel didn't know about it and still
contained tags-resources relationships between them marked
as active.
This commit adds resetting the model when any of the other models
is reset.

NOTE: it would be nice to make sure that TagsResources isn't reset
when a new Tag or a new Resource is created, since it
doesn't have any tags-resource relationships yet.