Fix removing a tag in KisTagChooserWidget
Before this commit, tag would only be removed when it's generated
by Krita, which means readonly (because it doesn't have an actual
representation in the database, it's just for viewing).
This commit changes this behaviour to the correct one,
allowing removal of only tags existing in the database.

Before this commit, the first tag from the combobox would be
selected at the end of the function no matter if the tag could
be removed or not. This commit changes it to select the first tag
only if the current tag is getting removed.