Fix undeletion of the last removed tag
Before this commit, undeletion of the last removed tag
wasn't working. The option for this wouldn't appear
and there was no logic underneath as well.
This commit adds this behaviour back in place.

Note: you can only undelete one tag, the one that was last
deleted and only in the same session of Krita.

This commit also removes the "Clear the list of deleted tags"
because this button is not needed if the list consists of
only one tag maximum.