Remove unused functions from KisResourceTaggingManager
Before this commit, there were a few leftover functions
in KisResourceTaggingManager that would be called from
KisTagChooserWidget to handle some operations on tags.
Because now operations on tags are done directly on
the KisTagModel in KisTagChooserWidget, there is no need
for KisTagChooserWidget to call functions from
KisResourceTaggingManager. This commit removes them.