Add filtering by tags in DbExplorer
Before this commit, no proper tag model would be created, assigned
and updated when the resource type changed. This resulted in no
filtering by tags in the Db Explorer in the Resource View tab.
This commit fxies this behaviour - it now gets the proper tag model
from the provider every time the resource type changes. The filter proxy
also now gets the current tag to filter the resources.