Make TagChooserWidget restore last tag automatically
Before this commit, after the reset of the model the current index
would have to be called manually every time something changes
in the model.
This commit makes sure it always comes back to the last used tag,
or 'All' in case the last used tag is missing (i.e. was just removed).