Make sure KisTagsResourcesModel can be reset
Before this commit, there was a bit of confusion regarding
prepareQuery and resetQuery. This commit removes that confusion
and uses resetQuery every time the data changes.
The function resetQuery doesn't do anything except for sending
signals. This is intentional because some other components
like filters now can use those signals to be invalidated.