Fix saving tasksets in new version
Before this commit, databasedbcache would look for tasksets
in 'tasksets' directory while the 'taskset' directory was
created. Also if the created directory was 'tasksets'
but no other changes were made, that would break the
compatibility with resource folders created in older versions of Krita
(tasksets made in older versions would be lost).
This commit ensures there is a correct directory being created
and that if there is a directory in the older path, it's renamed
so tasksets are loaded in the new version as well.