Fix assigning to a new tag in ItemChooserContextMenu
Before this commit, right-clicking on the item in the item chooser
(for example brush preset in Brushes docker), selecting "Assign to tag"
and putting a name for a new tag would result in creation of a new tag,
but no presets would be assigned to this tag.
This commit fixes this behaviour, now creating a new tag and
assigning it to the selected resource works.