Fix hiccups when doing canvas actions
Fixed KisSignalCompressor became too precise and tries to keep
the interval too hard. Sometimes the signal handler is too slow, in
such cases we should still give some time for event loop to execute.

To achieve that, KisSignalCompressor now has an additional operational
mode: "additive interval mode". When this mode is active, then interval
time is counted from the moment, when the execution path is returned from
the signal handler. In "precise interval mode" (default), in reverse, the
interval is counted from the moment, when start() signal arrives.

BUG:414576,415773
CC:kimageshop@kde.org