Fix tag being present in Remove from tag list
Before this commit, because of faulty use of std::remove,
which doesn't actually remove the tag, but only pushes it
at the end, tags weren't removed from lists. This commit
makes sure that both Assign to tag and Remove from tag
lists don't contain the current tag since it is in
"Remove from this tag" action.