Android: Add renderer in combobox
Prior to `4707a23f`, `supportedRenderers` would contain multiple renderers
which would activate combobox. But when we removed `RendererDesktopGL`,
it would only contain `RendererOpenGLES` which was until now, only
available for Windows.

The result of this being: when OpenGLES was disabled, there was no
possible way to activate it again on Android.