Fix threading issue in KisShapeLayer::transform()
The transformation itself should happen in the GUI thread, not in the
worker thread. Therefore we should use Qt::BlockingQueuedConnection.
It is not entirely safe (it can theoretically cause deadlocks), but it
is the only way how we can resolve the issue with keeping GUI thread the
only writer to the shapes.

The patch also fixes the problem in KisShapeLayerCanvas::forceRepaint() and
KisShapeLayerCanvas::forceRepaintWithHiddenAreas(), which should also
initiate the update in the GUI thread, not in the worker thread.

CCBUG:417016