Create a many-to-many table to link tags and storages
This also means that tags are independently active from
storages; deactivating all storage that contain a certain
tag does not hide that tag from the user. If a user tags
a resource from a storage that doesn't contain that tag,
there is no new link made from that storage to that tag.

This table only collects where the tags came from.