Fix possible deadlocks in KisShapeLayer and KisShapeLayerCanvas
Using Qt::BlockingQueuedConnection is dangerous. The GUI thread may
easily block on KisImage::barrierLock()/waitForDone() and the worker
thread will become blocked in the deadlock.

To overcome this issue KisBusyWaitBroker class was introduced. It
tracks all the locks of KisImage and breaks the tie of the deadlock
in favour of the worker thread, when needed.

The code that runs in a context of a worker thread and needs to pass a
blocking signal should utilize KisSafeBlockingQueueConnectionProxy class.
It passes the signal in a safe way, breaking the tie with KisBusyWaitBroker
if needed.

BUG:417233