Add "All Untagged" fake row in KisTagModel
From the discussion on the forum:
https://krita-artists.org/t/how-do-you-use-tags/1792
it turned out to be beneficial to add a new fake tag
called "All untagged" specifically because a lot of bundles
or a new brushes are not tagged at all,
or brushes imported as brushes, not from a bundle.

It would be good to have it hidden when there is no brushes
that fit under this category, but that might cause some issues
with synchronization. For now it's always in the tags combobox.