Implement favouritePresetsNamesList
Before this commit, the whole code in the KisFavouriteResourceManager
would use m_favouritePresetsList to get presets: both images and names
of presets to be loaded.
This commit is the first one changing it: this one changes the code
using KisFavouriteResourceManager to use a list of names
instead of a list of presets.