Fix dizzy outline in rotational brushes
We shouldn't modify actual 'pos' of KisPaintInformation, because
this position is used in calculation of drawingAngle(). Instead, we
should pass current zoom level to KisCurrentOutlineFetcher, so it could
align the outline properly.

Now there is a special methid KisAlgebra2D::alignForZoom() that aligns
he point onto a view-grid using the absolute zoom-level.