Disable liblzma (.xz) support for KArchive
KArchive links to LZMA incorrectly on Windows. It links to liblzma-5.dll,
even though we have only liblzma.dll. Since we don't use .xz functionality
of KArchive, we can just disable it on Windows.

The only place where we use LZMA is NSIS installer, we don't use
it in KArchive actually.