Remove use of resources() from SvgSymbolCollectionDocker
Before this commit, SvgSymbolCollectionDocker would use
a list of resources taken from the server. It means that all
of the resources used would be loaded at the same time, at opening the
docker. This commit changes it so only names are loaded.

Changes in this commit:
- instead of list of collections, KisResourceModel is used
- SvgCollectionModels are cached so there is no need to recreate
collection models (don't mistake them with the KisResourceModel)
- cache is cleared every time the KisResourceModel is reset
- in case of resource model reset, the id of the previously
selected collection is saved to set it up after the reset