Implement 'filter by tag' and delete 'update tag'
Before this commit, there was an 'update tag' button at the end of the
filter box under the resource dockers (for example Brush Presets
docker). This commit removes this button.
This commit also adds a 'filter by tag' button at the end of the filter
box, which gives you this behaviour:
- when checked, always shows only brushes from the tag
- when unchecked and the filter box is empty, still shows the resources
assigned to the tag (more sane behaviour)
- when unchecked and the filter box is not empty, shows all the
resources that conforms to the filter (no matter what tags they have
assigned)