Use the model to rename sessions
Before this commit, KisSessionManagerDialog would manually rename,
update to database and save to the file the resource.
This commit changes it to just calling the model to rename the
resource.
Note: in time of writing, renaming of the sessions doesn't work yet
because of the duplication of the resource type in the filename.