Add an assert in KisAslStorage
Before this commit, in some specific circumstances KisAslStorage
would crash in getting a pattern that is null even though it is saved in
the list of patterns (or a null layer styles from the list of layer
styles).
This commit adds an assert so if it happens for some ASL files, we'll
know about it. It's not a safe assert since I cannot guarantee it's safe
to return a null resource.