Disable loading anything from invalid asl file
Before this commit, invalid asl files with references to nonexistent
patterns would add their own patterns to the list. Because of the
resource system rewrite, it can no longer work because if there is a
dummy pattern added, there is still no actual file gthat we can later
use to load this pattern (since before it existed just in the memory,
it was fine back then).
This commit just marks the ASL as invalid if there are nonexistent
patterns references and the ASL Storage appears to be empty.