Fix updating resources in the folder storage
Resources were created with the resourcetype as part of the filename,
which is wrong; the filename of the resource should just be the filename.

The versioning of the filename of the resources has changed too, we
now use a suffix separated with a . from the main filename to do the
counting. This means that on saving, a resource with a dot in the name
will have the dot changed to an underscore in the filename (same as
with spaces in the name).