Implement KoEphemeralResource resource type
This type of resources has no representation on disk. They are created
on the fly by the corresponding factories. E.g. KisAutoBrush. These
resources don't have any load/save() methods, so they are implemented in
a base, KoEphemeralResource.

The main reason of this patch is to remove code duplication in
implementations of save()/load() calls and move actual implementation
to KoResource class.