Fix incorrect common curve in very old presets
Before this commit, some very old presets would get an incorrect
curve with a weird two hills characteristics, which is a specific curve
chosen to point out the inconsistencies in curves options code.
This commit fixes that by checking if the curve has been set and if not,
setting a clean empty curve instead.

BUG:417748