Fix crash on access null res in assert in KisLocator
Before this commit, the safe assert checking the filename of the
resource was placed above the assert checking the existing of the
resource. This commit adds a check so in case of a null resource
it is not written to the cache and the condition for a safe assert
for the filename is not checked.