Fix retrieving a pattern from an AslStorage
Before this commit, retrieving a pattern from the AslStorage
wouldn't work because it had a resourceType in the filename.
Now it removes the resourceType from the file path.
Note that in bundles and folders patterns are inside a folder,
while ASL is a file (not even an archive like a bundle), so
there is no folders inside, so the "path" is fake and the only
information we need is the uuid and if it's a pattern or a style.

Btw: since we do have the prefix in the file path, maybe we could
skip adding the _[resource type] at the end of the file path.