Remove KoResource::operator=()
It was implemented incorrectly: no operators in derived classes called
corresponding implementations of base classes, which basically caused
"slicing". It was easier to remove all operator=() implementations instead
of just trying to fix all of them. Anyway, they were never used (and that
is the reason the bug has never been seen).