Fix creating a database with an incorrect brush preset
Before this commit, Krita would by default crash on the first startup
(with no database or no resource folder). That was caused by a brush
preset that cannot be loaded anymore because it was based on an old,
deleted brush engine (Dyna brush engine). This commit fixes it by
checking the loaded resource for null.