Implement a class to set numeric HS<X> values for KisVisualColorSelector
Once attached to a selector, it will sync the values and switch
its labels according to the color model in use.
Note that for non-HSX models, it will be inactive and should be
disabled or hidden by the UI embedding it.

(cherry picked from commit fc92fa45aac3d841d9f4d5490a25d8809fad1066)