Add linked resources to the bundle
Before this commit, only resources requested by the user would be added.
This commit makes sure all linked resources are also added (in the time
of saving) so the bundle will be valid with no broken resources.

Note: because of the way bundle gets resources from the database based
on the manifest, for now there is no guarantee the correct resource will
be chosen because it's all based on md5, so filenames might be different
if the user happens to have multiple resources with the same md5.