Opacity Keyframes Cache Correction -- Affected Frames Correction
Opacity Keyframes highlighted an issue with the original algorithm for
detecting the number of affected frames with scalar keyframes. It wouldn't
always report the correct number of frames to calculate for cahcing. This
fixes that issue by taking into account the type of interpolation that
exists between the two keys.

Opacity Keyframes work with KisAnimationCachePopulator but have data-race
issues with the KisAsyncAnimationRenderBase. This is addressed in the next
commit.