Fix Palette Docker to refresh correctly after a color is added from color picker
The palette docker does not refresh correctly after any color
is added through color picker. Don't know if this is the right
approach to solve the problem but have solved it by setting
PaletteDocker as an observer for PaletteResourceServer and triggering
an update signal whenever the palette in the palette server is
updated using color picker.

BUG:418282