Change the TabletView into ListView in BundleManager
Before this commit, the BundleManager had TableView
to show all the storages. This commit changes it into a ListView
with an ItemDelegate that will be more covenient to use.

Note: the item delegate is incomplete and wrong for now.
It doesn't show the items it should because there is a mix of
indeces between the source model and the proxy model.