Fix KoFlake::resizeShape to work with null-width shapes properly
KoFlake::resizeShape() used to use absoluteToRelative mapping to save
the position of the still point. The problem is that this transformation
is not defined for zero-width shapes with still point **outside** the
line of the shape. absoluteToRelative maps the still point to this
line, basically losing information about distance of the still point
to this line.

To solve this problem we should include this still point transformation
into the main shape transformation procedure. We tried to be lazy and
avoid that, but it didn't work :)

(cherry picked from commit 8add377650992aa844fd9a2653bc83f4177ad153)