Attempt to add checkboxes in Bundle Manager
Before this commit, the item delegate in bundle manager
would not react at all to switching the selected item etc.
This commit adds changing the background to the proper selected
color and adds checkboxes to the items to be able easily
switch the activeness of them. The checkboxes don't work properly
for now.