Add just another warning message for incorrect ffmpeg
Before this commit, if the user downloaded the source code archive
(which is the huge green button on the ffmpeg.org website)
and tried to submit it as a ffmpeg path, it would just complain
about it being an incorrect file. This commit adds a simple explanation
why and what the user should do instead.