Do not load link profiles
Photoshop comes with at least two profiles that get installed
globally that cannot be used in Krita. AnimePalette.icc is one
of them. The type of this profile is not cmsSigNamedColorClass
which it probably should be, but cmsSigLinkClass, which I cannot
believe is correct.

A quick google shows that other applications also have trouble
with this broken profile.

BUG:419121