Do not load link profiles
Photoshop comes with at least two profiles that get installed
globally that cannot be used in Krita. AnimePalette.icc is one
of them. The type of this profile is not cmsSigNamedColorClass
which it probably should be, but cmsSigLinkClass, which I cannot
believe is correct.

A quick google shows that other applications also have trouble
with this broken profile.

BUG:419121
(cherry picked from commit 6e12760c0a2f0c29897d797ceefe593f28c47822)