When finding the screen at a point, use the midpoint of a widget's geometry
If you have two displays with different resolutions, say one is
1920x1200 and the other 1920x1080, then QWidget::geometry().topLeft()
might actually not be in a screen, so we try harder, by checking the
midpoint of the geometry. It would've been handy if QRect::midPoint
existed, but it doesn't.

So the deprecation of QApplication::desktop()->screenGeometry(),
now leads to code like

    QRect geom = parentWidget()->geometry();
    QPoint p(geom.width() / 2 + geom.left(), geom.height() / 2 + geom.top());
    QScreen *s = qApp->screenAt(p);

And we still need to check whether screen isn't 0...

(cherry picked from commit e2c286f4089c2871c7c0d980dcc97a29a6972ae3)