Fix "Preserve Lightness" to work with all color spaces
Now filling of the brush with color and applying alpha mask has been
moved into KoColorSpace as a single operation. It should be much faster
and it makes the code more understandable.

This patch also changes the "preserve lightness" algorithm. Instead of lerp
it uses standard overlay operation in the lightness scale. This approach
generates result a bit more like Ramon's brushes.