Modify the thumbnail of a new watercolor brush preset
This brush is a blending-only brush, by convention, the default
brush kit give a feedback to the user this brush doesn't "paint"
by using the combo of:
- A white stroke of painting
- Little blue drops of water as emblem
This commit fix the icon that previously wasn't intuitive for that.

Note: Blending only brush usually belong to the K) part of the
brushkit; but I think it is nice to keep this brush along the
collection of new brushes because of its specialised task for
watercolor.

(cherry picked from commit 9072ce2708351c1891b4a723d06b851d4d91f751)