Implement Fill Tool and magic wand on color labeled layers
Before this commit, Fill Tool and Contiguous Selection Tool
would have only two modes: to work on
a specific layer or to work on a projection of the whole image.
This commit adds a new mode: working on a set of layers that are
labeled with a specific color label. The user can choose it in
Tool Options.

Three modes:
- current layer
- all layers
- color labeled layers - there can be multiple labels selected


(cherry picked from commit 7748d8d18487dd02795cfcd379c226bd0653779b)