Implement Fill Tool on color labeled layers
Before this commit, Fill Tool would have only two modes: to work on
a specific layer or to work on a projection of the whole image.
This commit adds a new mode: working on a set of layers that are
labeled with a specific color label. The user can choose it in
Tool Options.