Disable "Color" and "Lightness" brush tip modes in engines not supporting them
These modes are supported only by "Pixel" and "Tangent-Normal" brushes.
It would be nice to implement that for "ColorSmudge" paintop.

WARNING: some old preset might have had "Color" mode for colorsmudge.
It was a bug in previous versions of Krita. This switch changed the brush
tip a little bit. Now this mode is not supported, all old presets will
automatically switch into Mask mode (there are no such preset in the
default pack of presets).

CC:kimageshop@kde.org