Timeline Docker Redux: Scrollbar Realignemnt and Zoom Feat.
This patch gives the timline docker the ability to zoom while scrolling.
It also improves some of the behavior regarding the infinite scroll by
adding an overscroll signal and overriding some of the default scrollbar
input handling. It also realigns the scrollbar to account for the header
offsets.

This should help timeline navigation and UI/UX problems. It will also
stop users from getting stuck while trying to scroll.

This can eventually be reused in the curves docker too...