Text Filter Now Accounts for Direct Parent Name
There are cases where I believe the text filter should work where either
masks are bound to a node of a specific name or there are direct
children of a group nome with a name that matches the filter. These are
now accounted for with an additional boolean. Maybe this could be
configured if users don't like it?