Limit cursor drag in mirror mode elliptic selector shape
Without any kind of barrier between left and right half,
it was not really possible to select the min or max value
as you could in a bar control.
With this change, the cursor coordinate will be limited
to the side where mouse dragging started.

(cherry picked from commit 418ebf108553dcd037eb01aba837c78257bca276)