Ensure that selection masks get unactivated if the selection was emptied.
This can happen with selectall and then invert, or selecting an area and then deselecting
it fully, which ended up with an empty but active selection, and thus an inability to paint
on the drawing.

Global selection mask will be removed, local selection masks disabled, and in either case
the artist can go back to painting.

BUG:420338