Krita 5: New Animation Timeline Docker.
This commit turns the Timeline Docker into the Animation Timeline
Docker, which will serve as Krita's new hub for traditional animation
and combines the functionality of the Animation Docker and Timeline Docker.

We've also created a new KisTransportControls widget as a simple,
reusable element for easily adding transport controls (play/pause, stop, prev
and next buttons) wherever necessary. With this comes the addition of
new pause and full stop functionality.

Designed and programmed with Eoin.
Review and corrections by Dmitry, Scott and Ivan. Thanks again!

BUG:377112
BUG:412907
BUG:377111