Remove 500ms delay between selection creation and cretion of marching ants
This delay was originally introduced to not interfere with painting,
which (in 2013) was done directly over the ants. Now we have overlay
mask functionality, therefore this delay is not needed anymore.

There is still a short visible delay. It is cause but the outline
genration algorithm is really slow. Perhaps we should also try to
fix speed when working on bug 419240.

BUG:419738
CCBUG:395861