Adjust the popup palette selector on display renderer changes
When the color space of a document or current layer changes,
the color selector did not react to this.

After a lot of brain racking I came to the conclusion that the
most sane solution is to make color space changes explicit
for KisVisualColorSelector, and pass it the original
foreground color again.
Changing color space was added to KisColorSelectorInterface too.

More implicit behavior and assumptions about expected outcome
seem counterproductive at this point.