Animation Timeline Docker: Settings Menu Widget.
While we've moved various settings that are important to the animation
workflow to the Animation Timeline Docker's titlebar, in order to cut
down on user interface clutter we've taken the remaining, set-and-forget
style options and hidden them in a new menu widget at the right side of
the Animation Timeline Docker's new titlebar.

OSX: Added macros for including types. (Thanks Ivan!)

Added localizable tooltips.
Added "Drop Frames" to the Animation Timeline.