Fix the fix of Brush HUD
When the property list was empty, the widget got way too small.
So better not bypass the top level layout completely, but set a
minimum size for the scroll area, and wait some more until the
top level layout recalculated a recommended size.

Things would be easier if QLayout::sizeConstraint had separate
horizontal and vertical parameters, using SetFixedSize would
resize before repaint but then we lose control of the height...