Fix memory leaks in KisMergeLabeledLayersCommand
Before this commit, KisMergeLabeledLayersCommand would keep the image
and paint device for as long as it is kept in the undo stack.
It's not needed, because the undo uses different ways to determine how
to undo and redo the selection.
After this commit at the end of the merge the command will clear its
shared pointer, which will in turn ensure that those objects will be
removed when nothing else will be using them.