Make the color labels combobox nicer in Tool Options
Before this commit, color labels combobox would look exactly like
in the filter label for layers, but stretched horizontally.
It means the view-filter icon when no filtering is used,
and a small circle with colors when there are colors selected.
This commit changes the icon to "tag", which seems more appropriate,
and the visuals of the combobox to be in simple rectangles instead of
a circle, so it fills the combobox better.